+++ pkg/awsglue/glue_table.go
@@ -42,42 +42,7 @@ const (
 	GlueTimestampType = "timestamp" // type in GlueTableMetadata tables for timestamps that we will re-map Go times
 )
 
-// Use this to tag the time partitioning used in a GlueTableMetadata table
-type GlueTableTimebin int
 
-const (
-	GlueTableMonthly GlueTableTimebin = iota + 1
-	GlueTableDaily
-	GlueTableHourly
-)
-
-func (tb GlueTableTimebin) Validate() (err error) {
-	switch tb {
-	case GlueTableHourly, GlueTableDaily, GlueTableMonthly:
-		return
-	default:
-		err = fmt.Errorf("unknown GlueTableMetadata table time bin: %d", tb)
-	}
-	return
-}
-
-// return the next time interval
-func (tb GlueTableTimebin) Next(t time.Time) (next time.Time) {
-	switch tb {
-	case GlueTableHourly:
-		return t.Add(time.Hour).Truncate(time.Hour)
-	case GlueTableDaily:
-		return t.Add(time.Hour * 24).Truncate(time.Hour * 24)
-	case GlueTableMonthly:
-		// loop a day at a time until the month changes
-		currentMonth := t.Month()
-		for next = t.Add(time.Hour * 24).Truncate(time.Hour * 24); next.Month() == currentMonth; next = next.Add(time.Hour * 24) {
-		}
-		return next
-	default:
-		panic(fmt.Sprintf("unknown GlueTableMetadata table time bin: %d", tb))
-	}
-}
 
 // Meta data about GlueTableMetadata table over parser data written to S3
 // NOTE: this struct has all accessor behind functions to allow a lazy evaluation
